refacto + Adding scenario promo noel + livraison France et Italie simples

+++ Dojo.BDD/IPromotion.cs
@@ -4,6 +4,6 @@ namespace Dojo.BDD
 {
     public interface IPromotion
     {
-        int CalculerRemise(Panier panier, Dictionary<TypeDeFruit, int> PrixFruits);
+        decimal CalculerRemise(Panier panier, Dictionary<TypeDeFruit, decimal> PrixFruits);
     }
 }
